use unix timestamp if version=0

+++ torrent.go
@@ -33,7 +33,7 @@ type Metainfo struct {
 
 	// Fields proposed by Herry et.al. (see DOMINO workshop paper)
 	UUID    string `bencode:"uuid,omitempty"`
-	Version int    `bencode:"version,omitempty"`
+	Version uint64 `bencode:"version,omitempty"`
 }
 
 // Signature holds data signature
@@ -45,7 +45,7 @@ type Signature struct {
 }
 
 // NewMetainfo creates a new Metainfo instance (torrent file) of given 'filePath'.
-func NewMetainfo(filename, uuid string, ver int, tracker string,
+func NewMetainfo(filename, uuid string, ver uint64, tracker string,
 	pieceLength int64, privkey *openssl.PrivateKey) (*Metainfo, error) {
 	mi := Metainfo{
 		UUID:         uuid,
